+++ data_engineering/pipeline_runner.py
@@ -2,7 +2,7 @@ import pandas as pd
 
 from utils import feature_imputation, pipeline_io
 
-from data_engineering.utils.derive import (
+from data_engineering.utils.other_derivation import (
     derive_total_assets_from_log_total_assets,
     derive_short_term_securities_from_cash,
     rename_column,
